Correccion en diseño de la barra de navegación desbordada y mejora en diseño del dashboard
