+++ frontend/src/components/Navbar.jsx
@@ -25,8 +25,6 @@ const Navbar = () => {
   const location = useLocation(); // Hook para obtener la ubicación actual
   const [showProfileMenu, setShowProfileMenu] = useState(false);
   const profileMenuRef = useRef(null);
-  const [showReportMenu, setShowReportMenu] = useState(false);
-  const reportMenuRef = useRef(null);
   // Función para determinar si un enlace está activo
   const isActive = (path) => {
     // Para rutas exactas
@@ -47,12 +45,13 @@ const Navbar = () => {
   useEffect(() => {
     const handleClickOutside = (event) => {
       if (
-        reportMenuRef.current &&
-        !reportMenuRef.current.contains(event.target)
+        profileMenuRef.current &&
+        !profileMenuRef.current.contains(event.target)
       ) {
-        setShowReportMenu(false);
+        setShowProfileMenu(false);
       }
     };
+
     document.addEventListener("mousedown", handleClickOutside);
     return () => {
       document.removeEventListener("mousedown", handleClickOutside);
@@ -97,22 +96,6 @@ const Navbar = () => {
     navigate("/home");
   };
 
-  // Cerrar el menú de perfil cuando se hace clic fuera de él
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (
-        profileMenuRef.current &&
-        !profileMenuRef.current.contains(event.target)
-      ) {
-        setShowProfileMenu(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
   if (!usuario) {
     return (
       <nav className="navbar-ae">
@@ -246,66 +229,6 @@ const Navbar = () => {
                 </span>
                 <span>MVA</span>
               </Link>
-              <div
-                className="nav-link-item-container"
-                ref={reportMenuRef}
-                style={{ position: "relative", display: "inline-block" }}
-              >
-                <div
-                  className={`nav-link-item ${isActive("/admin/reportes")}`}
-                  onClick={() => setShowReportMenu((prev) => !prev)}
-                  style={{ cursor: "pointer" }}
-                >
-                  <span className="nav-link-icon">
-                    <FileText size={18} />
-                  </span>
-                  <span>Reportes</span>
-                </div>
-                {showReportMenu && (
-                  <div
-                    className="report-dropdown-menu"
-                    style={{
-                      position: "absolute",
-                      top: "120%",
-                      left: 0,
-                      background: "#fff",
-                      boxShadow: "0 4px 18px #0002",
-                      borderRadius: "10px",
-                      zIndex: 10,
-                      minWidth: "170px",
-                    }}
-                  >
-                    <Link
-                      to="/admin/reportes-evento"
-                      className="dropdown-link"
-                      style={{
-                        display: "block",
-                        padding: "12px 18px",
-                        color: "#8a1538",
-                        fontWeight: 600,
-                        textDecoration: "none",
-                      }}
-                      onClick={() => setShowReportMenu(false)}
-                    >
-                      Por Evento
-                    </Link>
-                    <Link
-                      to="/admin/reportes-mes"
-                      className="dropdown-link"
-                      style={{
-                        display: "block",
-                        padding: "12px 18px",
-                        color: "#8a1538",
-                        fontWeight: 600,
-                        textDecoration: "none",
-                      }}
-                      onClick={() => setShowReportMenu(false)}
-                    >
-                      Por Mes
-                    </Link>
-                  </div>
-                )}
-              </div>
             </>
           )}
         </div>
