+++ frontend/src/views/admin/reportes/styles/AdminReporteDetalle.css
