--- frontend/src/views/admin/styles/AdminDashboardNew.css
@@ -1,196 +1,0 @@
-/* Dashboard styles */
-
-/* Variables de colores institucionales */
-:root {
-  --primary-color: #8a1538;
-  --secondary-color: #2c3e50;
-  --accent-color: #3498db;
-  --light-bg: #f4f6fb;
-  --white: #ffffff;
-  --text-dark: #2c3e50;
-  --text-light: #7f8c8d;
-  --shadow-soft: 0 10px 30px rgba(0, 0, 0, 0.08);
-  --shadow-strong: 0 15px 35px rgba(0, 0, 0, 0.1);
-  --gradient-bg: linear-gradient(135deg, #f4f6fb 0%, #e3e8f0 100%);
-}
-
-/* Estilos base del dashboard */
-.admin-dashboard {
-  background: var(--gradient-bg);
-  min-height: calc(100vh - 60px);
-  padding: 2rem;
-  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-}
-
-/* Encabezado del dashboard */
-.dashboard-header {
-  text-align: center;
-  margin-bottom: 3rem;
-  padding-bottom: 1.5rem;
-  position: relative;
-}
-
-.dashboard-header h2 {
-  color: var(--primary-color);
-  font-size: 2.5rem;
-  font-weight: 700;
-  margin-bottom: 0.5rem;
-  letter-spacing: -0.5px;
-}
-
-.dashboard-header::after {
-  content: "";
-  position: absolute;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 100px;
-  height: 4px;
-  background: linear-gradient(
-    to right,
-    var(--primary-color),
-    var(--accent-color)
-  );
-  border-radius: 2px;
-}
-
-.dashboard-subtitle {
-  color: var(--text-light);
-  font-size: 1.1rem;
-  font-weight: 400;
-}
-
-/* Contenido del dashboard */
-.dashboard-content {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1.5rem;
-  justify-content: center;
-}
-
-/* Panel de información */
-.dashboard-info-panel {
-  background: var(--white);
-  border-radius: 1rem;
-  box-shadow: var(--shadow-soft);
-  padding: 2rem;
-  text-align: center;
-  width: 100%;
-  max-width: 900px;
-  transition: transform 0.3s ease, box-shadow 0.3s ease;
-}
-
-.dashboard-info-panel:hover {
-  transform: translateY(-5px);
-  box-shadow: var(--shadow-strong);
-}
-
-.dashboard-info-panel h3 {
-  color: var(--primary-color);
-  font-size: 1.8rem;
-  margin-bottom: 1rem;
-  position: relative;
-  display: inline-block;
-}
-
-.dashboard-info-panel h3::after {
-  content: "";
-  position: absolute;
-  left: 50%;
-  bottom: -10px;
-  transform: translateX(-50%);
-  width: 80px;
-  height: 3px;
-  background: var(--primary-color);
-  border-radius: 2px;
-}
-
-.dashboard-info-panel p {
-  color: var(--text-dark);
-  font-size: 1.1rem;
-  line-height: 1.6;
-  margin-bottom: 2rem;
-}
-
-/* Espacio para futuros reportes */
-.dashboard-placeholder {
-  background-color: var(--light-bg);
-  border-radius: 0.75rem;
-  padding: 2rem;
-  margin: 2rem 0;
-  position: relative;
-}
-
-.dashboard-placeholder h4 {
-  color: var(--secondary-color);
-  font-size: 1.4rem;
-  margin-bottom: 1rem;
-}
-
-.dashboard-placeholder p {
-  color: var(--text-light);
-  font-size: 1rem;
-}
-
-.placeholder-icon {
-  width: 80px;
-  height: 80px;
-  margin: 0 auto 1.5rem;
-  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='24' height='24' viewBox='0 0 24 24' fill='none' stroke='%238A1538' stroke-width='2' stroke-linecap='round' stroke-linejoin='round'%3E%3Cline x1='18' y1='20' x2='18' y2='10'%3E%3C/line%3E%3Cline x1='12' y1='20' x2='12' y2='4'%3E%3C/line%3E%3Cline x1='6' y1='20' x2='6' y2='14'%3E%3C/line%3E%3C/svg%3E");
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: 50px;
-  opacity: 0.7;
-}
-
-.dashboard-note {
-  background-color: rgba(138, 21, 56, 0.08);
-  border-left: 4px solid var(--primary-color);
-  padding: 1rem 1.5rem;
-  border-radius: 0 0.5rem 0.5rem 0;
-  margin-top: 2rem;
-}
-
-.dashboard-note p {
-  color: var(--primary-color);
-  font-size: 0.95rem;
-  font-weight: 500;
-  margin: 0;
-}
-
-/* Responsive Design */
-@media (max-width: 768px) {
-  .dashboard-header h2 {
-    font-size: 2rem;
-  }
-
-  .admin-dashboard {
-    padding: 1.5rem;
-  }
-
-  .dashboard-info-panel {
-    padding: 1.5rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .dashboard-header h2 {
-    font-size: 1.8rem;
-  }
-
-  .dashboard-header {
-    margin-bottom: 2rem;
-  }
-
-  .admin-dashboard {
-    padding: 1rem;
-  }
-
-  .dashboard-info-panel h3 {
-    font-size: 1.5rem;
-  }
-
-  .dashboard-placeholder {
-    padding: 1.5rem;
-  }
-}
